requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.30